NewStuffPage: Rework to be a loader-based thing, add connection checker
There is still an awfully common error that causes some kind of breakage
that makes most of our text go away. This is extremely annoying, and
seems to be due to something with network access and one bit of the
searching for a solution suggests that we are attempting to post an
event to a null thing, allegedly usually caused by attempting to call
deleteLater on a null object, and apparently also related to QNAM.
Extremely annoying.