StatusInfo: Minor layout wiggling, add placeholders for global fx preset/edit
Got sidetracked by the previous thing, but... this one's on the list,
and i think this feels like the correct place for the stuff to live in
general. However, we need to be able to actually operate on it as well,
which is likely to be a somewhat larger operation.