Segment: Add a restartClips property, and supporting infrastructure
The intent here is to let segments require that a clip be restarted from
the beginning when that segment is reached during playback. The logic
is, stop the current playback of that segment (if any such thing is
ongoing) and start the clip again. Small thing, kind of, but it allows
but some fun compositional stuff.