SongManager: Add Split In Half functionality
The new button sits between Add Before and Add After, and the logic is
that it will split the current segment in two, which will have a
combined length of the current segment, with any uneven duration given
to the original segment (because one of them need to have it). Both will
have the same clip selection, but in case a clip is set to restart, it
will not be set to restart on the new segment (which is added after the
current segment). This ensures that splitting a segment will have no
audible effect, until the user does other things with that new segment.