PartBarDelegate: Add "will play/stop soon" indicator
This is now properly visualises what is about to happen for a clip, if
something is about to happen to it (specifically, if it has been asked
to start, it will blink a play icon to indicate that, and if it is
supposed to stop and is currently playing, a stop icon will flash).
Feels pretty good, will need testing, visuals likely want polishing up,
but... yup, seems to be doing the thing.