stepseq: Add UI for pattern swing, subnote ratchet, and subnote probability
This also adds pages to the NoteSettings and StepSettings popups. You
may notice there is only one setting on page two (probability), which i
did to make sure all the ratchet settings were on the same page. This
gives us a couple of extra bits we can add to those popups if we want,
or we can reorganise things so probability ends up on page 3, with the
ratchet settings on page 2, just to give us that "last page has the
fewest entries" layout that's more common. But, this works, anyway.