stepsequencer: Rework the NoteSettings popup, removing need for StepSettings
This was originally started as an intention to just make NoteSettings
show up without the kind of delay that just stops you in your tracks
when doing stuff. It evolved into a "make that one thing do both jobs"
and it can now handle arbitrary step ranges inside a bar, which... makes
life much simpler. It also feels nicer in use. Needs some fine tuning,
but it's already much nicer. It also shows up instantly, instead of
taking upwards of a couple of seconds with lots of notes on a pattern,
which... not really cool, so, yeah, much better now.