RecordingPopup: Pile of layout and option picking logic (mostly midi, also general)
The recording functionality itself isn't done yet, but it's coming
together really nicely from a visual angle. Of course, the options on
the audio settings page aren't as designed yet, but that wasn't the
intention behind those controls anyway, so i've left those mostly alone
for now. The design puts a number of the selection options under a
single item (in particular source, since "record master" basically makes
the source selection irrelevant, so it might as well go in that same
list, instead of being its own option... and so on for the other options
in the design).