last selected indicator: update explicitly instead of binding
The old code ended up with us having a funny, incorrectly positioned
rectangle on startup, and controlling it directly like this results in
us... not having that problem. It could be done with other hacks as
well, but this one's clear enough, and matches what we're doing in a few
other places as well, so that'll do.