RecordingPopup: Add and remove hardware devices for midi source selection
Using the new model on MidiRouter, and a handy proxy model to filter out
things that aren't hardware devices, this allows people to pick a
specific hardware device to record events on. Not necessarily the most
vital bit of functionality, but the plumbing was also needed to make a
bunch of other stuff work properly, so hey, nice.

Also meant finding that whole "there's two device entries for midi
5-pin" thing, which was kind of silly and perhaps a touch weird.