RecordingPopup: Ensure selectedChannel/slot are in sync with reality
The logic here is that, when you open the dialog, we store the selected
channel and the slot for that channel in the dialog, so that we can,
firstly, change the source of the sound without changing its destination
at the same time, and also reset it when the dialog is closed again
following either no recording being done, or a successful recording.

This also handles the case where a user might close the dialog during
recording time, at which point we'll want to ensure the information
remains the way it's supposed to (so only apply the above logic when
there's no on-going recording).

Also fix a minor string issue