RecordingPopup: Handle record and stop more properly (also fix small ref error)
We still need to handle stopping the recording when playback is stopped
during recording and we are recording in midi mode. This /should/ be
handled implicitly by SequenceModel and PatternModel anyway, but it's
still better to have it be just a little more explicit, and at any rate
needs testing. So, that's tomorrow's task: Testing and such to make sure
midi live recording actually works properly now.