gui.controller: Handle zctrl value change signal, and add propagation stop
We now listen to changes made by zctrl, so we can ensure our UI is
always up to date when the underlying control value changes, which used
to be less of a problem, but which becomes a problem as we move forward
and reintroduce midi learn.

The propagation stop is, again, to ensure that we don't end up
propagating changes that aren't actually changes. While there's a stop
elsewhere, this also ensures we don't just... spit out change
notifications at random, and do it more deliberately, which is always
good.