ControlPage: Add midi learn popup and handling of the same
The popup is shown when the three vertical dots are pressed next to the
value on a controller. In the popup, there are two options: Forget the
existing learned cc (which simply clears the current one), and learn new
(which will restore any existing learned value if aborted). Learning a
new one will show a "please wiggle a knob" type popup as it waits,
including the ability to abort the learning attempt. In short, a
complete but minimal ui for learning purposes.

Future options would include the ability to explicitly pick a channel
and cc, instead of doing learning. We might also want the ability for a
learned cc value to be limited, or squeezed, but this is something we
can get to much further down the line (and probably wanting to be
introduced as part of any mod matrix style work).