sketchpad.song: only stop clips in the current "song"
Still a great many items, but there's no particular reason to try and
stop things in song which are... not the current song. So, adjust things
just a touch here. We still touch all the clips (just in case the scene
has changed in weird ways), but now we're only stopping 50 clips instead
of 500 when we do stopAllPlayback.