sketchpad.clip: Actually send along correct song when stopping
With the previous change (stopping only the current song) this is
effectively a null change at the moment, since we have no way in the UI
of changing the active song. However, since all the data is still there
and we might eventually decide on a way to change things again, well...
why not.