PartBarDelegate: Reduce the updating work for playfield changes
This moves the logic for handling the PlayfieldManager's update signals
out of the delegates, where it kind of logically should live, to where
we only get hit once. This should nicely reduce the amount of
roundtrips, which are costly, and also just not make any of the calls
in places where it doesn't matter (such as the song player page).
Functionality's identical, but should have a fair bit less impact this
way.