ChannelKeyZoneSetup: Polish up and adjust to a few new bits
This is now a dialog in its own right, and uses our new knob controlling
logic (also because the old logic was entirely lost during that
transition, but mainly because this is where that wants to live). Also
adjust the visuals for the active-voices visualisation for each sample
to match the diamonds used on the sample visualiser etc.