MixedChannelsViewBar: Add new sample style selection picker, adjust keyzoning logic
The new selection picker just shows the value and lets you see the
dialog, in a similar style to the Routing selector.

Keyzoning now re-adds the ability to do custom management of the keyzone
setup for samples, without unnecessarily complicating things for people
who don't want to bother with that.