ChannelWaveEditor: Adapt to new playback style logic, and fix a bug
The fixed bug here is that we previously did not handle the case where
the first thing to happen on startup was that there would be a sample
selected. Consequently, the selection would be incorrect in the clip
settings bar component on this page. The new property on ClipSettingsBar
to allow for manually setting that information does the trick here.

It still seems weirdly slow to show the page, which will need some
investigating.