ComboBox: Swap out SVGItem for a Rectangle
This isn't a core UI element, in the sense that it's a self contained
popup, so as long as it's consistent with its own look, that will work
fine.

Also noticing we've got a bunch of places which use the QQC2 combobox,
which really want to probably be using this one instead.