PageManager: Prepare our cache in a threaded manner, nicer bootlog
As far as i can tell this has functionally no effect on the bootup
speed, even if the loading output would suggest otherwise. The reason
there is that it records the same start time for all the pages, and
really only records the total duration once completed for them all, so
that the longest duration is essentially the total load time from
loading starts until it ends.