MixedChannelsviewBar: Add warning for when there's synth/fx on tracks that won't use them
This shows up as a small exclamation mark be-iconed button on the Synth
button on the Skechpad Track overview. It shows up when you are not in
Synth mode but have at least one synth engine on the track. Similarly,
it shows up when there are fx on the track, but you are in Sketch mode
(where the fx are not used). When clicked, it shows a dialog which
explains why the button is there, and what you might want to do. We have
no way to hide it other than that, and i frankly don't think that we
should have one. However, if we need to be able to hide it, we can work
out how to do that.