ASDRClipView: Support A, D, and R values higher than 2 seconds
This switches from using the standard slider to using our own
InfinitySlider, which allows for functionally infinite values in an easy
to use and graceful manner. This means we can now have a functionally
infinitely long value for both attack, decay, and release. Release is
still limited, of course, but uses the infinityslider in its fully bound
mode because it looks nicer that all four are the same. The
interaction's the same anyway, so why not.