sequencer: Swap tap-to-select default to select-step from select-subnote
The swich control (alt) still applies, and you can select a specific
subnote directly by holding alt and tapping that subnote. It is,
however, excessively awkward to be quite that precise, so it makes sense
for that to be the "you're doing something very deliberate" action,
which kind of is what we probably want to do for anything that requires
a modifier.