sketchpad.channel: all-notes-off for the track, not arbitrary zynthian channel
This was at one point correct, but no longer really makes sense.
Instead, what we really want to happen, is that if there was something
there before, to ensure things aren't all mixed up, send out an
all-notes-off for the entire track, instead of targeting a specific
channel (which isn't really how we do that anyway).