stepsequencer: Fix erroneous clearing of listened notes with mpe controllers
Also rejig the logic just a touch, but the main purpose there is to
ensure we actually filter correctly for all-notes-off, instead of just
assuming it's the only thing where that byte is 123. This doesn't
usually show much, but it's particularly common on things like a
seaboard block, which, being my main controller for testing purposes, is
a little awkward. Useful for finding things like this, though ;)