stepsequencer: More clear select/deselect workflow
When you have a step or note selected, tapping the clear button will
always clear that selection fully.

When you have a step selected, that is shown in the sidebar (perhaps we
want to show the number of notes on that step, but also just showing the
step value is kinda nice).