Move edit_contextual handling to main
This way we are able to handle it centrally from where ever we happen to
be (we previously were locked out from e.g. the playground, as that
didn't handle the button itself). We really want to be able to switch
from anywhere, as well as just... having just a bit more heuristics in
there if we've not got anything selected.