stepsequencer: Don't sniff those last straggler notes when stopping playback
Prior to this, we'd end up regularly picking up the last couple of notes
sent out, as they're reported through the router's ui output signal in a
delayed fashion (deliberately, for playback performance reasons), so to
make sure we don't do anything with those, just wait a few moments after
playback has stopped and then handle things there. This isn't
necessarily a solution for all such problems (perhaps it'd make better
sense to actually send out a flag with the signal about whether it
happened during playback or not), but also, it works.