Rename the sample downloader to sketch downloader (as that's what it is)
It points at the Sketches category, and it works explicitly on Sketches.
We currently will take whatever is uploaded to that location, but it is
supposed to only be sketches, so...

It also leaves us open to potentially reintroducing a sample dowloader
which actually does that at some point in the future, if we want to.