Adjust the clear fx passthrough values for empty slots logic a bit
The previous fix introduced the possibility that if the pan was the
default, then we would not reset dryWetMixAmount. Also a small
adjustment to where we decide that value, since it really only needs to
be done the once, so we can do that particular check outside the for
loop (just a bit easier to read).