Song Manager: Follow the current playback segment during playback
This is arguably a little "move stuff under the finger" ish, but...
also, useful, and more useful than that one particular corner. Weighing
up things is hard sometimes, but it's more useful than it is potentially
dangerous (which, really, it just isn't anyway, at most it's three taps
to change the thing back).