SlotsBar/FilePicker: Switch Save As... to Save A Copy... for Sketches and Samples
Technically what we're doing is more akin to an export, but that just
confuses the terminology more than it really has to, and Save A Copy
describes more properly what we're doing. This avoids us confusing the
Save As term, because that isn't what we're doing here (unlike for
Sketchpads, where Save As... actually /does/ change the data location of
the thing we are going to be working on in the future, unlike Samples
and Sketches, where we keep working on the internal copy and just, well,
save a copy of the current state).