Fix replacing FX
This issue with replacing FX came up since adding the replace FX dialog and somehow it was overlooked for quite some time.
Old zynthian logic was using a variable `add_layer_eng` instead of directly passing which engine to start. Before adding the fx replace dialog, it was instantly replacing the existing fx with the new selected one. When the replace dialog was added, the fx is not replaced until the user presses the Yes button. During that time, somehow the `self.add_layer_eng` in zynthian_gui_layer was getting reset and hence the problem with replacing fx.
This commit removes the dependency on the variable and instead directly pass the engine to the method as a parameter