Disable the secondary option for spawning layer options selector
This dialog is ancient and functionally irrelevant at this point. The
only function in there that would make sense in our current state would
be the keyzone setup stuff, but that'd then be a single option in a
popup, and really it'd want to live in the same place it does for the
samples as well, so... just let's don't.