RecordingPopup: Further adaptations to new pattern APIs
This allows us to actually select the durations (step and pattern length
respectively) the same way it's done in the sequencer. It also adjusts
the logic for positioning the pattern duration in real time on the
thumbnail so that it actually lines up properly. Finally, there's now a
throttle, so that while you are changing the pattern length, we don't
end up stuttering visually. It's entirely possible we'll have to adjust
this, but ten ms seems reasonable enough as a throttle here, most of the
time the pattern wouldn't be updating that fast while recording (it's
possible, just not exactly common).