sketchpad_clip : Do not add `.sketch.wav` suffix to random samples when copying because they are not sketches.
Sketches can be unbounced but random samples are not generated from zynthbox and hence should not be allowed to unbounce. If a random sample is selected, it will have the metadata stored in them so that they can be restored on any zynthbox, but since they lack sound data, they cannot technically be called a sketch.