Sketch/TrackUbouncer: Handle the weird sample-loop situation when unbouncing
When unbouncing something set to sample-loop, we need to handle that
kind of gracefully. We can deduce that if this was recorded while the
track was in sample-loop mode, the recorded sounds would necessarily be
synth sounds (routing is to synth, explicitly, rather than the sampler),
so we can safely assume that's what happened there.