sketchpad.song: Add save/restore for the filter/compressor bits on passthroughs
This adds persistence for all the various types of clients, including
the ones we're not actually exposing in the UI yet. We will want to
consider where and how to expose those, but with this we can just focus
on that when it comes that time, rather than having to remember to also
them implement that.