ClipSettingsBar: Allow sub-semitone pitch shifting intervals
When pitch shifting a sample, it would often be for tuning purposes, and
that is regularly going to need to be pretty exact. So, let's allow it
to two decimal points worth of precision.

Also allow the full range of pitch shifting of +/- 48 semitones. It's a
bit weird at the outer edges, but weird things happen in sound design,
so why not, right?