sketchpad.clip: Adapt metadata to new CAS API, and use it as the truth
The metadata class still handles all the saving and loading of metadata
to and from the files, but no longer acts as an intermediary between the
UI and ClipAudioSource. All data that is used by CAS (and SamplerSynth)
is now stored directly in the CAS object instance, and no longer has a
local cache version. This allows us to not have to do quite so much
variable juggling, and also means we know to only act directly with the
CAS objects in the UI when changing these values, instead of having to
pass through Python to do that.