UI: Adapt to new sketchpad.clip/CAS API
This adapts to the new API, where interacting with a CAS directly will
cause sketchpad.clip.metadata to save, without having to outright pass
through that.

This also adds in the Wavetable playback style (essentially just a
variant of the Looping style that shows and interacts with the start,
stop, and loop points differently, in a way that makes sense when
interacting with wavetable files).