WaveEditorBar: Fix up a whole pile of interaction issues, reintroduce zooming
This was brought on by the desire to make wavetable navigation more
"normal", so to speak - that is, wanting to have the wavetable window shown,
and have it automatically shown when you scroll through the wavetable.
Perhaps some tweaking can be done to that logic, but it works pretty
nicely. (perhaps we want less of the overflow area in wavetables, or
perhaps none at all... but much easier to tweak that now it's actually
a thing at all)

Pinch-zooming still needs some work, but this is a vast improvement over
the old stuff, so... might as well get it in, so it's not just sat on my
machine.