ChannelWaveEditor: Rework layouting and knob interaction a bit
This gives us a touch more space to work with (which is nice), and the
huge margins were kind of not how things look elsewhere anyway, so...
yeah, let's do that.

Also, just pass along the handling of the cuia to the wave editor when
that's what we're after anyway, rather than duplicating that logic here
(especially now it's so much more fancy).