WaveEditorBar: Better focus rect on scrollbar, fix wavetable touch navigation
The previous visualisation was just a focus coloured rectangle
overlaying the selected area, which obscured more than it helped. The
new logic fixes this by having a transparent rectangle with focus-rect
coloured borders around the area instead. Much nicer.

Wavetable touch navigation wasn't clamped to window size increments, and
also focused using the old double-width-of-the-window logic, when
wavetable display is supposed to just be the full window.