ClipVoicesSettings: Actually add the subvoice settings (also EQ)
Not yet persisting the settings, but the editor now works as intended.
The editor currently lacks knob indicators, but we'll get to that
shortly (same as persistence).

The EQ functionality kind of just lives here because it physically fits, not
because they particularly belongs here. Perhaps something to reconsider
a bit later (for example, i have this idea that maybe we want the button
panel to extend up all the way along the left hand side, so it's not so
squished, and then we could also have a single note pad for triggering
the thing in a testy fashion (for samples, and a play/pause button for
sketches).