FilePickerDialog: Bunch more interaction work, plus layout, plus new buttons
The interaction work intends to make it simpler to navigate the
filesystem (so we now for example remember which thing we just navigated
out of, and the button-enabled logic is considerably tighter locked down
now, and more declarative).

The new layout gives us a bit more space to work with in the sidebar (so
we can show more information there as we move into tagging and the
like).

The new buttons are a view switcher (for switching between tags and file
system browsing), a new folder button (for creating new folders inside
the thing), and three buttons for renaming, moving, and deleting
whatever is selected. These don't yet do anything, other than them being
enabled/disabled as relevant to the current selection and folder.