RecordingPopup: Move clear clip dlg to main (similar to pattern clear)
This gets rid of an annoying comment on startup about not being able to
assign a thing to a thing, and also works around the issue where putting
it further into the hierarchy ends up with the dialog being unable to
find a window to open in. Some kind of bug deep in there somewhere, but
this works, so... it works.