sketchpad_{song,channel} : Make sampleset json a part of sketchpadjson so that sampleset state is saved for each version individually.
This is to solve the previous problem of tracks having a centralized sample-bank.json irrespective of the sketchpad version and hence all sketchpad versions were forced to use the same sample-banks for a track across all versions.

This commit implements sampleset saving/storing to be done like sketches. Just like sketches, all samples for all versions are stored in the same dir and the state is maintained in the individual version `.sketchpad.json` file