MidiControllerSettings: Add a WIP version of the settings panel for midi controllers
This is as yet incomplete, but once complete, this will let you edit all
the various settings available to change on midi devices. This includes
things like the mpe knowledge we have of the devices, as well as the
input and output filters, and whether the device wants to have midi
clock or beat signals sent.