MidiControllerSettings: Add enough UI that we can set input filters
This is still work in progress, but we can now construct filters which
can send cuia events and/or rewrite the incoming midi message to
another midi message. Output filters will want to be next, as well as
persistence logic (both explicit save/load type things, and automatic
loading of stored states for known devices).