autoconnect: Switch (dis)connect logic to using libzynthbox' JackConnectionHandler
This code ensures that we are only going to make the changes that we
actually make when performing a bunch of connections and disconnections.
In essence, instead of actually disconnecting and connecting ports, we
now only perform the final action on two ports, ensuring that we only
make the changes we actually want to end up with, rather than resetting
everything first. This, in turn, means that we no longer end up causing
unnecessary glitches when people change the sound engine setup during
playback.