Adapt to the Part -> Slot changes in libzynthbox
The Part concept was an outdated concept, from when we had such a thing
in the project. What we actually have in reality is an equal number of
slots on a sketchpad track for the types clip, sound source, and effect.
Clips are either a Sketch (looping audio) or a pattern (looping midi
data), a sound source can be either a sample or a synth engine
(controlled by the pattern, ignored when a track plays sketches or is
set to play to an external destination), and effects are the sound
effect processors that a sound slot's output can be routed through
(depending on track settings). While these types are kind of different
on a conceptual level, what they have in common is that they all fit
into a set of slots on a sketchpad track, and there is the same number
of slots for each of these. The reason for this is that we have a
concept by which a pattern plays only a single sound, which in turn is
routed to a single slot, and that requires there to be an equal amount
of all three of these types, and hence: Slot.