KnobIndicator: Slightly lignter, and simpler
The ridges are now only two rectangles, as opposed to eight, and we
don't add a border outline. This makes for a much simpler and more light
weight control, which is kind of important for something we want to use
all over the place. No huge impact on loading times and such, but even
then, no reason to be greedy. Also increased the rotation speed, because
it was functionally impossible to see before, so it's much more obvious
now.