DefaultEditPage: Fix fx-only tracks, add bypass control to sidebar, toggle with select
The first ensures that if we are on a track where there is no synth
engine defined, we still get shown the settings for the selected FX
engine (this is particularly important on any non-synth track, and just
generally was a bit silly, and likely from a long-ago type situation).

The second adds a toggle underneath the page list, bottom left hand side
of the page, which will show when the selected engine has a bypass
controller defined. This currently is true only for the jucy engines,
but as it now is a thing we can do at all, it'll be worth adding it in
for all of them.

There is perhaps an argument to be made that what we really want to do
is literally just pull out those controllers, and perform the bypass
operation on the passthrough clients, as we've kind of got some mixed
things going on with our wet/dry stuff, but... also that seems likely to
cause all manner of weird issues, so let's probably just not do that.

The last bit just means you can now press select (that is, yes, or bk)
to toggle bypass on and off whenever we're on this page, since select
wasn't used for anything else.