Rework our CUIA layout for higher consistency, plus some adaptation for that
This is not a complete rework, and likely will need further attention,
however it is already a solid step along the way.

It changes how many of our buttons are handled (so they now have a "pressed"
and "released" signal, allowing for some tighter control over how we interact
with the buttons). It also removes the "long" logic, leaving us with only short
and bold. Long has been long unused, and functionally bold is unused as well
for practically everything, and it may well make sense to get rid of that, but
that is something we'll need to look at a little more closely.

The intention here is to end up with the ability to detect whether any of the
keys are held down from anywhere any other key is handled (essentially
allowing any key to function as a modifier). This in itself will require more
work (that is, introducing the "ignore next release" logic from Mode in more
places), but it's a step in that direction.

Further, we introduce some more solid handling of the mini playgrid drawer's
display logic, which is now done more deliberately using a property.

A bunch of things called channel has been renamed to track, and a bunch of
things which had weird names for their cuia have been made more consistent
(for example, the word order of the mode button's cuia was wrong).

This also introduces logic for setting the filter cutoff and resonance via
CUIA, for all slots which have the setting available. NB: They are not yet fed
back correctly (see todo statement).

The cuia handling function has a rewriter for legacy calls at the top, which
allows us to handle the old entries as they come in from outside (if, say,
someone were to run the software using settings for a different piece of
hardware, such as a zynthian v5(.x) box, or just an old setup for our own).

Finally, a lot of debug spammery has been disabled, as it was kind of getting
a bit noisy.