python main: Add mixer display control logic (also, alt+screen1 toggles)
This requires qml stuff to go along with it, but the idea here is to
allow control of displaying the track mixer in an immediate form, by
holding alt and tapping the sketchpad button (under-screen button 1).

Doing so when already on the sketchpad page will toggle the mixer, and
doing so from any other page will switch to the sketchpad page, and then
force display the mixer.