Sketchpad main: Adapt to new mixer display logic, also fix up cuia callback
The larger part here is the cuia callback fix-up, but that should be
hopefully not detectable (other than less errors, that code was kind of
sus before).

The smaller, but more clearly detectable, part of this commit is the
adaptation to the mixer logic described in the parent commit of this
one.