python main: Add explicit handling of ignoreNextXButtonPress to cuia
This means we no longer need to handle this quite as explicitly when
using the various pressed buttons as modifiers. All that is now required
is to set the ignoreNext(Name)ButtonPress property, and the next release
of that button will not cause any action to be taken

We also introduce the ability to cancel those same actions, by pressing
the back/no button whenever one such button is held down. This allows
you to change your mind on things like playback control, if you've
pressed down a button, and realise you don't want to do that thing, and
you can then simply press the back button. Perhaps we want to also then
not have the back button itself do anything...