Recording popup/gui.sketchpad: Lightly rework post-recording track-type switching logic
If the recording popup is closed, we will do the following:
- if recording into a sample slot, switch to sample-trig mode
- if not, we recorded a sketch, so switch to sample-loop mode

If the recording popup is open, we will do the following when closing
it, if and only if the recording was not cleared prior to closing:
- if we recorded to a sample slot, switch to sample-trig mode
- otherwise we recorded into a sketch, so switch to sample-loop

Note: We will not switch track mode when closing the recording popup if
there was no recording performed, *or* if we cleared the most recent
recording before closing. The idea here is to ensure that the user will
end up looking at the thing they just created, or if they aborted the
creation, they will see the thing they started with.