Add color to sketchpad tracks, and use it (and allow setting it)
This introduces the concept that a track with no sound on it should be
grey, and any other track should be given a specific colour. It is an
incomplete implementation, as colours in other parts of our ui
(including the hardware interface's led colours) are using the old
logic, but this is a good beginning.