theme_chooser: Adjust default track colours some
Took a little bit of exploration, but we need colours that are visually
distinct both on the display, and when dimmed down for use as button
backlight colours in a dimmed-down state. This is likely not the final
set, but it should work for now. In particular, there's no blues in
there, as we use that for indications, and that just makes things
awkward.