python main: do_long_task now takes an optional message, and starts cb faster
Instead of having to remember to manually set the task message, we now
instead allow you to just pass that message along to do_long_task
instead.

Thinking the old logic waiting for ages before starting the task was
kind of... based on perhaps a misunderstanding of some of the various
bits of logic, as it really just sat there doing nothing for ages before
it started doing the thing it'd been asked to do (thinking the original
intent was to start the task, but not *show* the loading screen before
then... which is wonky logic anyway and just never worked like that,
and could have full well been written by yours truly a while back).