sketchpad.clip: Switch to ujson for the clip json operations
according to benchmarks[1], ujson is several times faster than json, and
is a drop-in replacement where there are no special encoding things to
worry about, such as the clip data here. We probably couldn't use this
anywhere there's user-input strings to worry about, but here it's fine,
and fast is good for us. If it turns out to be a real problem, we can
always switch to orjson, but... would rather we didn't have to do that
until we have to ;)