python main: Fix master volume by using the new dryGainHandler client
Previously the volume was just a linear assignment of gain from 0
through 100, which in itself was never correct according to our mixer's
visuals anyway, but it also just was incorrect audibly as well, so...
this fixes that problem, and all our gain adjustment bits are now much
more correct, yay!

There are likely more spots that'll want some adjustment, but this fixes
the core issue.