file prop helper: Expose preview clip as a property, don't auto-loop preview
The automatic looping goes counter to how we want to actually use the
functionality (that is, as a kind of trig-at-root-note style preview),
and so we now instead just run playback once.

We also expose the playback CAS object itself, so we can use that to do
such fun things as displaying the position of playback and the like.