Sample Library: Previews, load-to-slot, better knob nav, layout fixups
We now have the ability to preview samples as we navigate through the
list (currently just by tapping the little button on the sample, but we
probably want that to live somewhere a little more prominent, for ease
of access)

Tapping an already selected sample will let you load it into a slot
(also if you click select on the sample again). The currently selected
slot will be automatically pre-selected in the popup.

We now only show the folders which might actually potentially have
samples in them (that is, not the root of the category, as that should
never have anything in it, and while you technically can, we should
discourage that).

The slot entries in the leftmost column now look more like they should
(label with useful information on empty slots, and the background for
the waveform even when no sample is loaded).