sketchpad.channel: Introduce externalSettings property, deprecate old ones
The externalSettings property contains an object that holds all the
various settings for external mode for a track, including the logic
that'll hold the settings for the selected module (so we don't end up
with modules eating each other's settings, which would be sucky).

The old ones are still used around the place, so leaving them in place
just for now while i work on the new functionality, but that'll want
removing at some point.