Add a default External module, with 4x3 knobby controls
The 12 controls can be interacted with the same way you would those on
the default synth edit page, except here you can also change which midi
channel, and which CC control the knob is going to operate on.

It currently will always start at cc value 0, but thinking we likely
want to do some funkiness so we store and restore that between sessions?
Need to work on that a bit more.