knsrc: Prepare the configs for the new knewstuff functionality
This puts the data into subdirs named for the author. This ensures that
if the author did something silly with the naming (like clash with
others), then at least it's only a problem for them, not for others.

Also, we need to have a QuestionAsker implementation sorted out, so that
we can actually forward those questions the installer will ask (like
"hey, that file's there already, wanna overwrite?" type stuff).