sketchpad.channel: Add setChannelSoundFromSnapshotSlot function
This is a kind of partner function to the ones for setting sample and
sketch slots from a specific slot of a snapshot. The intent is to allow
us to pick a specific synth or fx, and apply that to a specific slot on
some track. Particularly useful for webconf, but i suspect also useful
for on-device stuff at a later time.