sketchpad.channel: Switch off the mixer adjustment for track type
This was first implemented to assist in more easily handling the weird
effect of bouncing something and then ending up with the fx applied
again. Since we now have fully separate FX chains for sketches, that is
no longer an issue, and so we can kind of ignore that. If we do decide
we want it for something, it'll be easy to return with this change, but
for now it'll be simply locked to "synth".