BouncePopup: Switch to using the segments' before/after timer commands
This ensures that even with wider clip recordings than a total of 32768
steps worth, we are still able to actually control our recording
desires. Ran into this during Friday's messing about with creating very
many sketches, and it resulted in a serious lack of confidence in the
functionality in general from myself, and didn't want it to hang around
for too long, so, off duty time for working this one out.