sketchpad.channel: Add new trustExternalDeviceChannels prop, default sample picking style to all
Also removes the old same-or-first option (previously default), which
was all manner of confusing, because it resulted in samples working
fundamentally differently to how the other sound slots on a track would
work. To make them work differently should be a conscious choice, and
this lets us do that.

trustExternalDeviceChannels will define whether, when in Same sample
picking mode, we trust whatever an external controller gives to us or
not. By default we do not (and instead use the currently selected clip
as the destination), but if the user wishes to, they can change that.