TracksBar etc: Redo the auto-slot-selection logic for startup
This ensures that our selected slots are actually the appropriate
selected slots, using the first and best logic, but only setting the
selected slot now, and *not* the last selectd object, which causes all
manner of problems if we do that. So, it now only happens on first
switch to a track that hasn't had that done yet, which should... be fine
i think?